message: Correct jsdoc on `Message#flags`: not actually in events.
Learned this in my investigation for reviewing #3029; confirmed
empirically, after I couldn't find evidence for such a property
in the server code.

Also explain that it *is* in fetch results -- also confirmed
empirically, plus that's what our code depends on in that case,
and empirically flags like `mentioned` do work there -- and,
since the previous commit, in our `EVENT_NEW_MESSAGE` actions.