webview js: Add polyfill for Element#closest.
(Pairing with @borisyankov.)

We use this in a couple of spots already; we'll use it in
a few more soon.

This isn't actually necessary by the supported-version cutoffs listed
above.  But we'd already written it before we gathered the data to
produce those cutoffs, and it's cheap to keep around.