Fix a performance issue with some lists
Closes #3103

Both FlatList and SectionList offer as the main performance benefit
rendering a 'window' and not their whole contents. Sometimes in the
past, by mistake, we have rendered one of these in full height but
scrollable thanks to the ScrollView they are inside.

By analyzing all of our usage, I pinpointed the only major problem
we have, that resulted in a performance issues in the Emoji picker
and the `UsersScreen` (when one searches the 'all users' list),
which manifested as the list updating too infrequently during search.

Our default `Screen` behavior is to be 'scrollable' by default.
We should consider changing the default behavior. In the meantime
any screen that has a `FlatList` or `SectionList` directly inside
should switch to `contentScrollable={false}`.