android notif: Inline some spurious indirections from the wix library.
This is a pure refactor.  Instead of calling
 * the `onReceived` provided by the library,
 * which in turn calls `postNotification(Integer)`
 * which calls `buildNotification`
 * and then `postNotification(Notification, Integer)`
 * which calls `postNotification(int, Notification)`...

... just write down the five lines of code that those work out to.
(It'd be six, except we get to skip calling `notifyReceivedToJS`,
because our JS code doesn't subscribe to the event that that sends.)

Did I mention that that stack of five different calls includes three
which have the same name -- two of which have nearly the same
arguments, but in the opposite order -- and the three of them spread
out across the source file, with 9 other function definitions between?

Anyway, glad to be a little less entangled with this library's code.
This also gives us more control to make changes to this logic.

From pairing with @borisyankov.