emoji: Straighten out emoji-reaction types.
The type we called SlimEventReaction doesn't describe anything we
actually see at runtime, and nor does the generality of `user: any` in
EventReaction.  Instead, both reaction events and reactions in message
fetch results look a lot like what we have in EventReactionCommon.

So, pull that out and call it Reaction.  Replace uses of EventReaction
and SlimEventReaction with that.  Because it's part of the server API,
put it in apiTypes.js.

This is a pure type-annotations change.