api: Abandon `console.warn`, just use `console.log`.
In ebda24a9b "api: Fix noisy console.error on API error responses",
we switched from passing details on a failed API request to
`console.error`, to passing them to `console.warn`.

The point of that was to avoid a red-screen error in development when
a perfectly normal API failure happens.  But now we seem to be getting
those again (perhaps after the RN v0.57 upgrade?), with this error:

    argument[0].indexOf is not a function

Apparently RN's `console.warn` is implemented in such a way that it
expects a parameter with an `indexOf` property (eg: string/array).
We pass an object for the sake of nice inspectability in the Chrome
developer console.

We don't actually make use of these messages being `warn` rather than
`log`: we've had `console.disableYellowBox = true` since 57120ba83,
upgrading to RN v0.54, back in March, so they don't produce
yellow-boxes and only show up in the debug console, much like anything
else.  And `log` works fine with an object.  So, use `console.log`.

[greg: edited commit message]