message render: Refactor to group some properties logically.
The goal of this series is to (a) get a handle on the sprawling list
of "named parameters" / properties we're passing to a lot of different
functions in this subsystem, and (b) cut down the number of places
we're abusing the type of MessageList's `props` for other uses that
make it hard to track just what data is being relied on where.

This first commit just takes some of these properties, the ones that
are used in rendering messages but aren't specific to any particular
message, and groups them as one object to be passed down the stack.