title: Condense jsdoc on `getTitle*Color` selectors.
The parts about where this might be called aren't necessary here, and
are the sort of information that's at high risk of becoming wrong as
the codebase changes.

The part about how these selectors don't depend on the nav state is
important in explaining the contrast between this version and previous
versions which did depend on that -- which makes it important
information to include in the commit message where we made that
change, and we did include it there.  It's not important information
for understanding the current code in its own terms, so take it out
from the actual source code.