screen: Get `canGoBack` as a prop, instead of from Redux nav state.
In continuation to past few commits, this is a next step in removing
usage of deprecated `connectPreserveOnBackOption`.

We want to provide back-navigation in almost all cases, so make the
default `true` and mark only the exceptions.

For any normal navigation within the app, we want to offer navigation
back.  The only exceptions are screens the user can reach without
navigation, as the initial screen of the app; and when they've just
logged out, so can't go back to the logged-in screen they were on.
All not-normal-navigation cases appear in `navReducers.js`.

These boil down to:
 * `loading` -> LoadingScreen
 * `welcome` -> WelcomeScreen
 * `account` -> AccountPickScreen
 * `main` -> MainScreenWithTabs

Of these, two don't use `Screen` and already can't show a "back" UI,
so only `WelcomeScreen` and `AccountPickScreen` require attention.

[greg: reoriented analysis to clarify why no other screens need to
 pass this prop too.]