narrow: New abstraction caseNarrow for destructuring-bind of narrows.
This will give us a nice explicit, type-checked way of getting at the
bits of data inside one of our "narrow" data structures, depending in
the appropriate way on what kind of narrow it is.

As a demo, rewrite a couple of existing functions in this file this
way.  Note in particular:

 * In `isMessageInNarrow`, it's now unavoidably explicit that it
   returns `false` for a search narrow; without a `search` case,
   Flow will complain.  Previously it wasn't at all obvious
   whether there might be some type of narrow that didn't appear.

 * In `canSendToNarrow`, it's now explicit exactly which types of
   narrows produce `false`, as well as `true`.