StreamListCard: Remove a bunch of dead code.
The `StreamList` component hasn't taken a `clearInput` prop since
7107e6b70, last year; so stop passing it.  Then our `clearInput`
property isn't used.  Nor is `handleFilterChange`; it never has been.

As a result `setState` is never called, so the state is constant, and
we can take that code out too.

This was found by Flow after making StreamList.Props an exact type.