flow: Begin using some scattered exact types, uneventfully.
In general we want to make almost all our object types exact, with
narrow exceptions where there's a specific reason for an inexact
object type.

As an experiment and a start on this, I tried converting a sprinkling
of object types in various areas of our codebase to exact object
types.  A couple of cases that were a bit interesting were in the
preceding commits; this commit contains the rest, where there isn't
much to say.