message render: Tighten messageTagsAsHtml, and mark strict-local.
On setting strict-local, Flow complains "sketchy null check on number"
because `timeEdited` might be 0, which is falsy.  Fix that by
consistently actually testing for `undefined`; and by replacing
`?number` with `number | void` to reflect the fact that it really will
be only those and not `null`.

Also take the opportunity to just tighten up this code, eliminating
multiple overlapping checks on `timeEdited`.  This isn't *quite* a
pure refactor, in that the HTML output changes slightly: we get rid of
some extraneous whitespace.  That fixes a minor layout bug, where when
a message both has been edited and is starred, there was too much
space between the gray "tag" markers for those two facts.