tools: Fix files-changed helper to work on macOS.
This script, which we use to make our tests run by default only
on files that have changed, didn't work on macOS because it tried
to use `getopt` to parse its arguments, and macOS ships a `getopt`
command that is literally from the '90s:

  $ man getopt | tail -1
  BSD                       April 3, 1999                      BSD

and in fact was an exercise in nostalgia even then, harkening back
to... the '80s at latest, maybe '70s?

  $ man getopt | grep -B1 'Bell version'
  Written by Henry Spencer, working from a Bell Labs manual page.  Behavior
  believed identical to the Bell version.  Example changed in FreeBSD ver-

Kind of an obvious failure mode in retrospect, though; I should
have tested.

Fortunately this script doesn't really need any options, at least
not for now; just subcommands.  And those are super easy to parse
in pure shell when you have no options.  So just do that, and
rename some things appropriately.

Fixes #3179.