lint: Remove almost all indiscriminate `eslint-disable` markers.
With this, just one file remains where we disable ESLint without
scoping that down to a line, to specific rules, or both:

  $ rg 'eslint-disable [^a-z]'
  src/webview/html/processAlertWords.js
  4:/* eslint-disable */

That one has the excellent excuse that the bulk of it is copied
verbatim from the webapp, which is in a very different style of
JavaScript; and in fact our ESLint setup gives 25 errors and warnings
based on 10 different rules.

This change sweeps up the files that required no changes beyond
scoping the `eslint-disable` down to the spellchecker.  The previous
two commits handled more interesting cases, including one live bug.