howto/release: s/iTunes Connect/App Store Connect/g.
More or less.

Also move the Safari tip to "Troubleshooting".  It's been less bad
for me in recent weeks, and so I've stuck to Chrome.  (Though I think
this is unrelated to the name change; the webapp doesn't look any
different, so I don't think they coupled the name change with a big
software overhaul or anything.)