messageActionSheet: Stop using translated strings as identifiers.
It's already kind of a code smell to be using friendly user-facing
phrases like "Add a reaction" as identifiers internal to the code to
distinguish one callback from another.

One sign of where that can lead is how several phrases like "Unmute
stream" are duplicated in multiple places so that things will break if
they don't match up, and there's nothing to guarantee they match if we
go tweak one of them for wordsmithing's sake and forget to update the
others.

It gets worse when the identifier isn't even the string that's in this
code at all, but whatever *translated* string we might use in the
user's language.  Those are data that can, and ought to be able to,
change without notice, so they shouldn't affect the control flow we
have to reason about.

For a start, switch to the English versions as internal identifiers,
translating them only where we actually send them to be shown to the
user.  Subsequent steps will follow up.

This test was the one spot we used `extractString`, a tiny helper
function; cut it out.