messageActionSheet: Move handler metadata to property on function.
I think I like the locality of this arrangement better, with the title
right next to the actual action to be taken.  I definitely like that
the list of options in `allButtonsRaw` is now 100% boring as a list of
identifiers; having it be just *mostly* boring but also have important
information in it was not so good.