react: Stop repeating `props: Props;`.
We already mention the props type in the type argument to
`PureComponent` (or `Component`), and Flow is smart enough to
infer the obvious conclusion about the type of `this.props`;
so there's no need to repeat it.

RN upstream mostly eliminated this pattern at the same time as
adding that type argument to `Component` and friends, in some
commits in 2017 that mentioned upgrading to Flow v0.53.0.

Happily, this change was quite easy to make:

  $ sed -i '/props: Props;/d' -- $(rg -l 'props: Props;')
  $ tools/test lint --fix

(The second step got rid of a bunch of now-redundant blank lines.)