tools/upgrade: Get the RN deps into an in-program hash table.
This makes it much simpler to use various of these version numbers
for other purposes.

Though the calling convention for having a shell function do this
sure is absurd, and the `IFS=` trick for parsing is even more so.
It's possible this script should become my first experiment in
Node-based CLI scripting...