types: Make most types in `actionTypes` exact.
Done with the same command as in some previous commits:

  perl -i -0pe \
    's/^((?:export )?type \S+ = \{)(\n.*?\n)?(\};)/$1|$2|$3/msg' \
    src/**/*.js

but this time keeping the changes in src/actionTypes.js.  The
last several commits represent the various type problems that
this exposed; finish by actually making the types exact.

Also add some fixmes in tests, where (at least right now) it
isn't worth fleshing out these fixtures to make the types right.

There are a number of other types here that aren't yet covered,
because this command doesn't handle the `Foo & { ...` pattern;
that can be for a followup.