android notif: Inline a bunch of wix code to tell React notif opened.
From digestNotification, then from a bunch of one-line helpers for it
which are scattered through the base class.

That code conditionally does the InitialNotificationHolder thing in a
couple of places.  We've been unconditionally doing it up front, and
it's just setting a member on a singleton so it's idempotent -- so
just collapse those into doing it once up front.

Note these members mApp* and mJsIOHelper are themselves all various
classes from the wix library, so we're still calling a bunch of that
code.  But, in addition to being a step toward untangling all that,
with this change we are no longer calling any methods at all on the
PushNotification base class!  That brings us close to breaking the
subclass relationship, which will be another big reduction in our
surface area exposed to that library.