styles: Switch to getting static styles statically.
This in fact completely removes the use of contextStyles
from 22 of the 37 components that were using it.

This has the effect of changing back to `styles.foo` a large fraction
of the spots that the parent commit changed to `contextStyles.foo` --
with the difference that `styles` now means an imported module with
static data, whereas before the parent commit it meant an object
passed down through (the deprecated old React API for) context.

Performed the changes with these commands:

  names=$(perl -0ne 's/^\}.*//ms; print(join(" ", m/^  (\w*):/mg), "\n")' \
          src/styles/{misc,nav,utility}Styles.js)
  for name in $names; do
    perl -i -0pe '
      if (s(contextStyles\.('$name')\b)
           (styles.$1)g) {
        m(import styles,? )
          || s(.*import\K( \{.*\} from .\.\./styles.;))
              ( styles,$1)ms
          || s(.*\n import.*?\ from\ .*?\n \K)
              (import styles from '\''../styles'\'';\n)xms
      }' \
      src/*/*.js
  done

  perl -i -0pe 's(^.*contextStyles.*$)()m' \
    $(tools/test lint |& grep '\.js$')

  tools/test prettier --fix