settings: Identify each language with its own name for itself, too.
The names are from those used by Wikipedia in its own mobile UIs.
Wikipedia is a very conscientiously international and intercultural
project with a lot of effort going into it by speakers of many
languages, making it a useful gold standard for this.

One reference directly available on the Web is the list of Wikipedias:
  https://meta.wikimedia.org/wiki/List_of_Wikipedias

One option not found on that list is "Chinese (Simplified)" -- there's
just one Chinese Wikipedia for both simplified and traditional scripts:
  https://en.wikipedia.org/wiki/Chinese_Wikipedia
The translation we use does appear, though, in the "Language" box in
preferences in the Wikipedia webapp:
  https://en.wikipedia.org/wiki/Special:Preferences

[greg: added discussion in commit message]