android notif: Inline event name "notificationOpened".
Having a literal in the `emit` call is how the RN upstream docs
suggest doing it, and how almost all the RN upstream code does it.
It also seems clearly optimal short of a way to have the name
shared between the Java code and the JS code... which I haven't
yet seen done in the few bits of code that try to be abstract by
writing the name somewhere else.