events: Subscribe to all server events.
Omitting the `event_types` value subscribes the client to
all server events. Do that.

When I first decided to subscribe selectively only to events we can
handle it was the early days of the app and we could recognize just
a small part of all events.

Now we handle almost all events and the ones we don't we conveniently
log in the console.