android notif: Implement registering with GCM, stubbing the JS side.
This implements the entire GCM-related side of the work of registering
the app with GCM.  The RN-related side is left stubbed out; the next
commit will add that, and then we'll be able to remove the wix library
completely! 🥳

This is all nearly verbatim from the GCM upstream docs:
  https://developers.google.com/cloud-messaging/android/client

In particular from the sample app:
  https://github.com/googlesamples/google-services/tree/3e2518f6d/android/gcm/app/src/main/java/gcm/play/android/samples/com/gcmquickstart/
  https://github.com/googlesamples/google-services/blob/3e2518f6d/android/gcm/app/src/main/AndroidManifest.xml

The one bit that's at all specific to our app, or to RN apps in
general, is where we invoke RegistrationIntentService at app startup.
The sample app does this from the main activity's onCreate.  That's
pretty reasonable, but for an RN app it feels more natural to put it
in the `initialize` of a relevant React-module class.  (Which is just
a way of organizing the same thing -- `ReactActivity#onCreate` kicks
off many things, one of them `CatalystInstance#initialize`, which
calls `initialize` on each module.)  So do that.