nav: Ignore the empty rehydrated nav state.
This code existed to take us back, on relaunch, to the remembered
navigation state from a previous run of the app.

We changed that behavior long ago in favor of returning to the
main screen.  We discard the `nav` part of the state tree when
persisting the state, so this code should never do anything; if
it ever does anymore, that's a bug.  Remove it.

Also fix a bunch of tests that are still testing for the
old behavior.