account: Replace empty-string fallback with assert, in getCurrentRealm.
Much like the parent commit.  Each call site already depends on the
assumption that `realm` is a real realm URL and not the empty string.

At one call site, `WebLink`, it isn't directly obvious we have that
assumption, and in principle with certain ways of using it we
wouldn't; but we must assume it whenever `href` is relative, which
in fact it is in the one place we use `WebLink`.  (The semantics in
turn of `getFullUrl` are pretty broken, but fixing that will be for
another day.)