account: Add `tryGetValidAuth`, and use where appropriate.
These are remaining spots where we use `getAuth`, and don't assume
that there is some account/auth object there for it to return.

In both call sites, we're only interested in an auth object that
actually has an API key; so specify this selector accordingly.

For checking that an `Account` object really has an API key, we rely
on what the type tells us: the `apiKey` property is there, and it's a
string.  That's been true since forever for Account objects of
logged-in accounts, naturally, and since sometime in 2016 for
logged-out accounts; and since e0a32f93f last March for newly-added,
never-logged-in accounts.  That's well past the horizon of
invalid-data scenarios we're prepared to handle in general.