account: s/Valid//g in names; it's the invalid case that's special.
In general, when there's a value that might be a real, genuine Foo but
might also be an invalid Foo-like thing, it's best to scrupulously
label that fact in the names chosen in the source code.  (Unless the
context otherwise somehow makes this super clear.)

That's because inevitably when the name says it's a Foo, a person
reading the code will often take the name at its word.  (If someone
doesn't, and assumes everything might be a lie until proven otherwise,
it'll take them a huge effort to read very much code at all.)

Historically we hadn't done that with `getAuth` -- its return value
might have been a genuine, valid, logged-in auth object (one that
actually carries authentication to a Zulip server), or despite the
name it might have been an invalid one that had a dummy value (the
empty string) in place of an API key.  So a few commits ago, when
introducing related functions that did deal only in valid auth
objects, we named them `tryGetValidAuth` and `hasValidAuth`, to make
the distinction explicit.

Now `getAuth` lives up to its name.  So name these companion functions
simply `tryGetAuth` and `hasAuth`, to match.